simple-research-agent/requirements.txt
langchain
wikipedia
langchain-community
langchain-openai
python-dotenv
pydantic
duckduckgo-search
langchain-core
langchain-classic
duckduckgo-search